requirements.txt
django-allauth==0.24.1
django-appconf==1.0.1
django-auditlog==0.3.2
django-bootstrap3==6.2.2
django-cors-headers==1.1.0
django-easy-pjax==1.2.0
django-extensions==1.5.7
django-filter==0.11.0
django-fsm==2.3.0
django-fsm-log==1.2.1
django-fsm-admin
django_admin_bootstrapped
django-grappelli==2.7.3
django-jsonfield==0.9.15
django-material==0.4.1
django-registration-redux==1.2
django-rest-auth==0.6.0
django-slowdown==0.0.1
djangorestframework==3.2.5
docopt==0.6.2
docutils==0.12
factory-boy==2.5.2
fastcache==1.0.2
django_tables2
djangorestframework-jwt